Resolve delete update array

+++ src/app/shared/components/cards-accueil/cards-accueil.component.ts
@@ -3,6 +3,7 @@ import { ArticlesService } from 'src/app/core/http/articles.service';
 import { Article } from 'src/app/shared/models/article.model';
 import { LoginService } from 'src/app/core/services/login.service';
 import { EditorService } from 'src/app/core/services/editor.service';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-cards-accueil',
@@ -13,11 +14,9 @@ export class CardsAccueilComponent implements OnInit {
 
   htmlStr: string;
   public isLogin = !this.service.isLogin();
-  articlesList: Article[] = [];
 
   @Input() public article: Article;
   @Input() public index;
-
   @Output()
   deleteCard: EventEmitter<any> = new EventEmitter();
 
@@ -25,6 +24,7 @@ export class CardsAccueilComponent implements OnInit {
     private articlesService: ArticlesService,
     private service: LoginService,
     private editorService: EditorService,
+    private toastr: ToastrService,
   ) { }
 
   ngOnInit() {}
@@ -40,12 +40,17 @@ export class CardsAccueilComponent implements OnInit {
     this.editorService.typeEdition = false;
   }
 
-  deleteActivity(id) {
-    this.articlesService.deleteArticle(id).subscribe(
-      (articles) => {
-        this.deleteCard.emit(id);
-      },
-    );
+  deleteActivity(id, index) {
+    const result = confirm('Voulez-vous vraiment supprimer cet article ?');
+    if (result) {
+      this.articlesService.deleteArticle(id).subscribe(
+        (articles) => {
+          this.toastr.success('Article supprimé');
+          this.deleteCard.emit(index + 1);
+        },
+      );
+    } else {
+      this.toastr.error('Article non supprimé');
+    }
   }
-
 }
